+++ Fivemoreminutes/app/src/main/java/fivemoreminutes/cs499/cs/csupomona/edu/fivemoreminutes/fragments/Group.java
@@ -1,27 +1,27 @@
 package fivemoreminutes.cs499.cs.csupomona.edu.fivemoreminutes.fragments;
 
-import android.content.Intent;
+import android.app.AlertDialog;
+import android.content.DialogInterface;
+import android.support.v4.app.FragmentTransaction;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.inputmethod.EditorInfo;
-import android.widget.AdapterView;
 import android.widget.EditText;
 import android.widget.ListView;
-import android.widget.TextView;
+
 
 import com.getbase.floatingactionbutton.FloatingActionButton;
 
 import java.util.ArrayList;
 
 import fivemoreminutes.cs499.cs.csupomona.edu.fivemoreminutes.R;
-import fivemoreminutes.cs499.cs.csupomona.edu.fivemoreminutes.activities.DrillDownAlarmActivity;
 import fivemoreminutes.cs499.cs.csupomona.edu.fivemoreminutes.adapters.GroupItemAdapter;
-import fivemoreminutes.cs499.cs.csupomona.edu.fivemoreminutes.data.AlarmItem;
 import fivemoreminutes.cs499.cs.csupomona.edu.fivemoreminutes.data.GroupItem;
+import fivemoreminutes.cs499.cs.csupomona.edu.fivemoreminutes.model.AddGroupTask;
+import fivemoreminutes.cs499.cs.csupomona.edu.fivemoreminutes.model.DBHandler;
+import fivemoreminutes.cs499.cs.csupomona.edu.fivemoreminutes.model.GetGroupsTask;
 
 /**
  * Created by Calvin on 4/11/2015.
@@ -29,10 +29,10 @@ import fivemoreminutes.cs499.cs.csupomona.edu.fivemoreminutes.data.GroupItem;
 // In this case, the fragment displays simple text based on the page
 public class Group extends Fragment {
 
-    private ArrayList<AlarmItem> test = new ArrayList<>();
-    private ArrayList<GroupItem> groupItems = new ArrayList<>();
+    private DBHandler dbHandler = new DBHandler(getActivity(), null, null, 1);
+    private ArrayList<GroupItem> groupItems = new ArrayList<GroupItem>();
     private GroupItemAdapter listAdapter;
-    private ListView listView;
+    private int count = 1;
 
     public static Group newInstance(int page) {
         Bundle args = new Bundle();
@@ -41,7 +41,6 @@ public class Group extends Fragment {
         return fragment;
     }
 
-
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -49,61 +48,50 @@ public class Group extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        final View view = inflater.inflate(R.layout.group_fragment, container, false);
-
+        View view = inflater.inflate(R.layout.group_fragment, container, false);
 
         FloatingActionButton fab = (FloatingActionButton) view.findViewById(R.id.group_fab);
-        fab.setOnClickListener(new View.OnClickListener() {
+        fab.setOnClickListener(new View.OnClickListener(){
             @Override
-            public void onClick(View v) {
+            public void onClick(View v){
+                AlertDialog.Builder alert = new AlertDialog.Builder(getActivity());
 
-                final EditText editText = (EditText) view.findViewById(R.id.group_name);
-                editText.setVisibility(View.VISIBLE);
-                editText.setOnEditorActionListener(new TextView.OnEditorActionListener() {
+                alert.setTitle("New Alarm Group");
+                alert.setMessage("Enter the Alarm Group's Name");
 
-                    @Override
-                    public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
-                        boolean handled = false;
-                        if (actionId == EditorInfo.IME_ACTION_DONE) {
-                            String groupName = editText.getText().toString();
+                // Set an EditText view to get user input
+                final EditText input = new EditText(getActivity());
+                alert.setView(input);
 
-                            if (!isGroup(groupName)) {
-                                groupItems.add(new GroupItem(groupName));
-                                listAdapter.notifyDataSetChanged();
-                            }
-                            handled = true;
-                            editText.setVisibility(View.INVISIBLE);
-                            editText.setText("");
-                        }
-                        return handled;
+                alert.setPositiveButton("Ok", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int whichButton) {
+                        String value = input.getText().toString();
+                        groupItems.add(new GroupItem(value));
+                        count++;
+                        listAdapter.notifyDataSetChanged();
+                        Object[] params = { getActivity(), groupItems, listAdapter, value };
+                        new AddGroupTask().execute(params);
                     }
                 });
+
+                alert.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int whichButton) {
+                        // Canceled.
+                    }
+                });
+
+                alert.show();
             }
         });
 
-        listView = (ListView) view.findViewById(R.id.lv_group);
+        ListView listView = (ListView) view.findViewById(R.id.lv_group);
         listAdapter = new GroupItemAdapter(this.getActivity(), groupItems);
-        listView.setClickable(true);
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                Intent intent = new Intent(getActivity(), DrillDownAlarmActivity.class);
-                getActivity().startActivity(intent);
-            }
-        });
         listView.setAdapter(listAdapter);
+
+        Object[] params = { getActivity(), groupItems, listAdapter };
+        new GetGroupsTask().execute(params);
+
         return view;
     }
-
-    private boolean isGroup(String groupName) {
-        //checks if a string is in the group arraylist.
-        for (int i = 0; i < groupItems.size(); i++) {
-            if (groupName.equals(groupItems.get(i).getName())) {
-                return true;
-            }
-        }
-        return false;
-    }
-
 
 }
